Calificacion DTO correccion

+++ correos-web/src/main/java/co/edu/uniandes/csw/correos/dtos/CalificacionDTO.java
@@ -92,7 +92,7 @@ public class CalificacionDTO {
     /**
      * @return la calificacion de las calificacion
      */
-    public double getCalificacion() {
+    public int getCalificacion() {
         return this.calificacion;
     }
